mcps/scada/pyproject.toml
[project]
name = "envest-scada-mcp"
version = "0.1.0"
description = "Multi-instance SCADA MCP server (stdio + streamable-http) with tool prefixing and ephemeral auth."
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.20.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "pypdf>=4.0.0",
  "pymysql>=1.1.1",
  "pytesseract>=0.3.10",
  "pypdfium2>=4.30.0",
  "uvicorn>=0.30.0",
  "starlette>=0.37.2",
  "jinja2>=3.1.4",
  "openpyxl>=3.1.0",
  "xhtml2pdf>=0.2.15",
  "python-docx>=1.1.0",
  "bcrypt>=4.0.0",
]

[project.scripts]
scada-mcp = "scada_mcp.cli:main"
scada-mcp-manager = "scada_mcp.manager.app:main"
scada-mcp-combined = "scada_mcp.combined:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scada_mcp"]
